+++ src/Profile/Shopware/Converter/ShippingMethodConverter.php
@@ -947,8 +947,8 @@ abstract class ShippingMethodConverter extends ShopwareConverter
             $value['fromTime'] = '00:00';
         } else {
             $secs = $ruleData['bind_time_from'];
-            $min = \str_pad((string) \floor(($secs % 3600) / 60), 2, '0', \STR_PAD_LEFT);
-            $hour = \str_pad((string) \floor(($secs % 86400) / 3600), 2, '0', \STR_PAD_LEFT);
+            $min = \mb_str_pad((string) \floor(($secs % 3600) / 60), 2, '0', \STR_PAD_LEFT);
+            $hour = \mb_str_pad((string) \floor(($secs % 86400) / 3600), 2, '0', \STR_PAD_LEFT);
             $value['fromTime'] = $hour . ':' . $min;
         }
 
@@ -956,8 +956,8 @@ abstract class ShippingMethodConverter extends ShopwareConverter
             $value['toTime'] = '00:00';
         } else {
             $secs = $ruleData['bind_time_to'];
-            $min = \str_pad((string) \floor(($secs % 3600) / 60), 2, '0', \STR_PAD_LEFT);
-            $hour = \str_pad((string) \floor(($secs % 86400) / 3600), 2, '0', \STR_PAD_LEFT);
+            $min = \mb_str_pad((string) \floor(($secs % 3600) / 60), 2, '0', \STR_PAD_LEFT);
+            $hour = \mb_str_pad((string) \floor(($secs % 86400) / 3600), 2, '0', \STR_PAD_LEFT);
             $value['toTime'] = $hour . ':' . $min;
         }